fix: 병합 오류 해결

+++ src/main/java/damon/backend/service/ReviewService.java
@@ -11,6 +11,7 @@ import damon.backend.repository.ReviewLikeRepository;
 import damon.backend.repository.ReviewRepository;
 import damon.backend.repository.user.UserRepository;
 import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
@@ -27,6 +28,7 @@ import java.util.Optional;
 import java.util.stream.Collectors;
 
 @Service
+@Slf4j
 @RequiredArgsConstructor
 @Transactional
 public class ReviewService {
@@ -39,7 +41,9 @@ public class ReviewService {
     private final CommentStructureOrganizer commentStructureOrganizer;
 
     // 등록
-    public ReviewResponse postReview(ReviewRequest request, List<MultipartFile> images, String identifier) {
+    @Transactional
+    public ReviewResponse postReview(ReviewRequest request, String identifier) {
+
         User user = userRepository.findByIdentifier(identifier).orElseThrow(UserNotFoundException::new);
 
         Review review = Review.create(request.getTitle(), request.getStartDate(), request.getEndDate(),
@@ -47,20 +51,23 @@ public class ReviewService {
                 request.getContent(), request.getTags(), user);
         review = reviewRepository.save(review);
 
-        // 여러 이미지 처리
-        if (request.getImages() != null && !request.getImages().isEmpty()) {
-            for (String imageUrl : request.getImages()) {
-                ReviewImage reviewImage = new ReviewImage(imageUrl, false, review);
+        if (request.getImageUrls() != null) {
+            for (String url : request.getImageUrls()) {
+                ReviewImage reviewImage = new ReviewImage(url, review);
                 reviewImageRepository.save(reviewImage);
                 review.addImage(reviewImage);
             }
         }
+
+        review = reviewRepository.save(review);  // 리뷰 엔티티 업데이트
+
         List<ReviewCommentResponse> emptyCommentsList = new ArrayList<>(); // 새 리뷰에는 댓글이 없으므로 빈 리스트 생성
         return ReviewResponse.from(review, emptyCommentsList); // 저장된 리뷰와 빈 댓글 목록을 전달
     }
 
     // 수정
-    public ReviewResponse updateReview(Long reviewId, ReviewRequest request, List<String> deleteImageUrls, String identifier) {
+    @Transactional
+    public ReviewResponse updateReview(Long reviewId, ReviewRequest request, String identifier) {
 
         Review review = reviewRepository.findById(reviewId)
                 .orElseThrow(ReviewNotFoundException::new);
@@ -76,53 +83,35 @@ public class ReviewService {
         review.update(request.getTitle(), request.getStartDate(), request.getEndDate(), request.getArea(),
                 request.getCost(), request.getSuggests(), request.getContent(), request.getTags());
 
-        // 이미지 추가 및 삭제
-//        deleteImages(deleteImageIds, review);
-//        addImages(images, review);
-//        review = reviewRepository.save(review);
-
-//        // 이미지 삭제
-//        if (!deleteImageUrls.isEmpty()) {
-//            deleteImageUrls.forEach(awsS3Service::deleteImageByUrl);
-//        }
-//
-//        // 이미지 추가
-//        if (newImages != null && !newImages.isEmpty()) {
-//            for (MultipartFile file : newImages) {
-//                try {
-//                    String url = awsS3Service.uploadImage(file);
-//                    ReviewImage image = new ReviewImage(url, false, review);
-//                    reviewImageRepository.save(image);
-//                } catch (IOException e) {
-//                    // 예외 처리
-//
-//                }
-//            }
-//        }
-        // Delete images
-        if (deleteImageUrls != null) {
+        // 이미지 삭제 로직
+        List<String> deleteImageUrls = request.getDeleteImageUrls();
+        if (deleteImageUrls != null && !deleteImageUrls.isEmpty()) {
+            log.info("리뷰 이미지 삭제: {}", deleteImageUrls);
             for (String url : deleteImageUrls) {
                 reviewImageRepository.findByUrlAndReview(url, review).ifPresent(image -> {
-                    awsS3Service.deleteImageByUrl(image.getUrl());
+                    log.info("데이터베이스에서 이미지 {} 삭제 중", url);
                     reviewImageRepository.delete(image);
+                    reviewImageRepository.flush(); // 변경 사항 즉시 반영
+                    log.info("데이터베이스에서 이미지 {} 삭제 완료, S3에서 삭제 요청 중", url);
+                    awsS3Service.deleteImageByUrl(url);
+                    log.info("S3에서 이미지 {} 삭제 요청 완료", url);
                 });
             }
         }
 
-//        // Add new images
-//        if (newImages != null && !newImages.isEmpty()) {
-//            for (MultipartFile file : newImages) {
-//                try {
-//                    String url = awsS3Service.uploadImage(file);
-//                    ReviewImage newImage = new ReviewImage(url, false, review);
-//                    reviewImageRepository.save(newImage);
-//                } catch (IOException e) {
-//                    throw new ImageCountExceededException();
-//                } catch (MaxUploadSizeExceededException e) {
-//                    throw new ImageSizeExceededException();
-//                }
-//            }
-//        }
+        // 새로운 이미지 URL 처리
+        List<String> newImageUrls = request.getNewImageUrls();
+        if (newImageUrls != null && !newImageUrls.isEmpty()) {
+            log.info("새 이미지 추가 중: {}", newImageUrls);
+            for (String url : newImageUrls) {
+                ReviewImage newImage = ReviewImage.createImage(url, review);
+                review.addImage(newImage); // 리뷰 객체에 이미지 추가
+                reviewImageRepository.save(newImage);
+                log.info("새 이미지 추가됨: {}", url);
+            }
+        }
+
+        reviewRepository.save(review); // 변경사항 저장
 
         // 댓글 구조를 다시 조직화
         List<ReviewCommentResponse> organizedComments = commentStructureOrganizer.organizeCommentStructure(reviewId);
@@ -132,24 +121,25 @@ public class ReviewService {
     }
 
     // 이미지 등록
-    public List<ReviewImage> postImage(Review review, List<MultipartFile> images) {
-        List<ReviewImage> savedImages = new ArrayList<>();
-        try {
-            if (images!= null && !images.isEmpty()) {
-                for (MultipartFile file : images) {
-                    String url = awsS3Service.uploadImage(file); // S3에 이미지 업로드
-                    ReviewImage newImage = ReviewImage.createImage(url, review); // 이미지 생성
-                    savedImages.add(reviewImageRepository.save(newImage)); // DB에 저장
+    @Transactional
+    public List<String> postImage(List<MultipartFile> images) {
+        List<String> imageUrls = new ArrayList<>();
+        if (images != null && !images.isEmpty()) {
+            log.info("Processing {} images", images.size());
+
+            for (MultipartFile file : images) {
+                try {
+                    String imageUrl = awsS3Service.uploadImage(file, "review/");
+                    imageUrls.add(imageUrl);
+                } catch (IOException e) {
+                    throw new ImageCountExceededException();
+                } catch (MaxUploadSizeExceededException e) {
+                    throw new ImageSizeExceededException();
                 }
             }
-            return savedImages; // 저장된 이미지 리스트 반환
-        } catch (IOException e) {
-            throw new ImageCountExceededException();
-        } catch (MaxUploadSizeExceededException e) {
-            throw new ImageSizeExceededException();
         }
+        return imageUrls;
     }
-
 
     // 상세 조회 (댓글 포함)
     @Transactional(readOnly = true)
@@ -202,9 +192,6 @@ public class ReviewService {
             reviewImageRepository.delete(reviewImage);
         }
 
-        // 리뷰와 연관된 모든 이미지를 S3에서 삭제
-//        review.getReviewImages().forEach(image -> awsS3Service.deleteImage(image.getFileKey()));
-
         reviewRepository.delete(review);
     }
 
@@ -248,48 +235,16 @@ public class ReviewService {
                 .collect(Collectors.toList());
     }
 
-    // 이미지 삭제
-//    private void deleteImages(List<Long> deleteImageIds, Review review) {
-//        review.getReviewImages().forEach(image -> {
-//            if (review.getReviewImages().contains(image)) {
-//                awsS3Service.deleteImage(image.getFileKey());  // 파일 키를 사용하여 S3 파일 삭제
-//                reviewImageRepository.delete(image);
-//            }
-//        });
-//    }
+    // 내 리뷰 조회
+    /*@Transactional(readOnly = true)
+    public Page<ReviewListResponse> searchMyReview(String identifier) {
+        User user = userRepository.findByIdentifier(identifier).orElseThrow(UserNotFoundException::new);
+        Pageable pageable = PageRequest.of(page, pageSize, Sort.by(Sort.Direction.DESC, "id"));
+        Page<Review> myReviews = reviewLikeRepository.findMyReviews(user, pageable);
 
-    // 이미지 추가
-//    private void addImages (List < MultipartFile > images, Review review){
-//        validateImages(images);
-//
-//        images.forEach(image -> {
-//            try {
-//                AwsS3Service.UploadResult uploadResult = awsS3Service.uploadImage(image);
-//                ReviewImage reviewImage = ReviewImage.createImage(uploadResult.getFileUrl(), uploadResult.getFileKey(), review);
-//                review.addImage(reviewImage);
-//                reviewImageRepository.save(reviewImage);
-//            }
-//
-//            // 이미지 업로드 후 예외
-//            catch (IOException e) {
-//                throw new ImageUploadFailedException();
-//            }
-//        });
-//    }
+        List<Review> myReviews = reviewRepository.findMyReviews(user.getId());
+        return myReviews.map(ReviewListResponse::from);
+    }
 
-    // 이미지 업로드 직전 검증 로직
-//    private void validateImages(List<MultipartFile> images) {
-//        final int MAX_IMAGE_COUNT = 10;
-//        final long MAX_IMAGE_SIZE = 10 * 1024 * 1024; // 10MB
-//
-//        if (images.size() > MAX_IMAGE_COUNT) {
-//            throw new ImageCountExceededException();
-//        }
-//
-//        for (MultipartFile image : images) {
-//            if (image.getSize() > MAX_IMAGE_SIZE) {
-//                throw new ImageSizeExceededException();
-//            }
-//        }
-
+     */
 }
